out_stackdriver: update helper functions
- change int to int64_t to avoid memory leak

+++ plugins/out_stackdriver/stackdriver_helper.c
@@ -52,7 +52,7 @@ void try_assign_subfield_bool(msgpack_object obj, int *subfield) {
     }
 }
 
-void try_assign_subfield_int(msgpack_object obj, int *subfield) {
+void try_assign_subfield_int(msgpack_object obj, int64_t *subfield) {
     if (obj.type == MSGPACK_OBJECT_STR) {
         *subfield = atoll(obj.via.str.ptr);
     }
